Fix span length calculation bugs.
git-svn-id: https://svn.apache.org/repos/asf/pivot/branches/1.5.x@955972 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk/src/org/apache/pivot/wtk/content/NumericSpinnerData.java
@@ -105,7 +105,7 @@ public class NumericSpinnerData implements List<Integer> {
             throw new IllegalArgumentException("Lower bound must be less than upper bound.");
         }
 
-        long length = ((upperBound - lowerBound) / increment) + 1;
+        long length = (((long)upperBound - (long)lowerBound) / increment) + 1;
 
         if (length > Integer.MAX_VALUE) {
             throw new IllegalArgumentException("Bounded range is too large.");
